ref: move methods to `GlobalRegistryCollection`
