+++ src/CoreLib/Game/Zone/Supervisors/ZoneObjectSupervisor.cs
@@ -56,55 +56,10 @@ internal sealed class ZoneObjectSupervisor(Core.Zone zone) : ZoneEntitySuperviso
             var requirements = objectInfo.m_spawnRequirements.m_requirements.ToList();
             var operatorType = objectInfo.m_spawnRequirements.m_operator;
             
-            return CheckGlobalRegistryRequirements(requirements, operatorType);
+            return GlobalRegistryCollection.CheckGlobalRegistryRequirements(requirements, operatorType);
         }
 
         return true;
-    }
-
-    private static bool CheckGlobalRegistryRequirements(List<Requirement> values, Requirement.Operator operatorType) {
-        var allMatched = true;
-
-        foreach (var requirement in values) {
-            if (requirement is ReqGlobalRegistryValue globalReq) {
-                if (!GlobalRegistryValueMet(globalReq)
-                    && operatorType == Requirement.Operator.ROP_AND) {
-                    return false;
-                }
-
-                allMatched = allMatched && !globalReq.m_applyNOT;
-            }
-            else {
-                Logger.Warning("Holy!!! We found a spawn requirement that isn't a global registry value. " +
-                            "This is a problem. Let Jooty know.");
-            }
-        }
-
-        return allMatched;
-    }
-
-    private static bool GlobalRegistryValueMet(ReqGlobalRegistryValue value) {
-        var globalValue = GlobalRegistryCollection.GetRegistryEntry(value.m_entryName);
-
-        switch (value.m_operatorType) {
-            case ReqNumeric.OPERATOR_TYPE.OPERATOR_EQUALS:
-                return value.m_numericValue == globalValue;
-            case ReqNumeric.OPERATOR_TYPE.OPERATOR_LESS_THAN:
-                return value.m_numericValue < globalValue;
-            case ReqNumeric.OPERATOR_TYPE.OPERATOR_LESS_THAN_EQ:
-                return value.m_numericValue <= globalValue;
-            case ReqNumeric.OPERATOR_TYPE.OPERATOR_GREATER_THAN:
-                return value.m_numericValue > globalValue;
-            case ReqNumeric.OPERATOR_TYPE.OPERATOR_GREATER_THAN_EQ:
-                return value.m_numericValue >= globalValue;
-            case ReqNumeric.OPERATOR_TYPE.OPERATOR_UNKNOWN:
-            default: {
-                    Logger.Error("Zone contains a spawn requirement that " +
-                                      "references a global registry value that does not exist. " +
-                                      "Entry name: {EntryName}", Logger.Args(value.m_entryName));
-                    return false;
-                }
-        }
     }
 
 }
