+++ src/CoreLib/WizardData/Collections/GlobalRegistryCollection.cs
@@ -3,25 +3,26 @@
  * Proprietary and confidential.
  */
 
+using System.Collections.Generic;
 using System.Linq;
 using Imlight.Common;
 using Imlight.CoreLib.WizardData.Databases;
 using Imlight.CoreLib.WizardData.Models;
 using Imlight.CoreLib.WizardData.Models.World;
 using Raven.Client.Documents;
+using static Imlight.Common.Caches.TypeCache;
 
 namespace Imlight.CoreLib.WizardData.Collections;
 
 public static class GlobalRegistryCollection {
+
     private const string CollectionName = "GlobalRegistry";
 
     private static readonly IDocumentStore s_store;
     private static bool s_isInitialized;
     private static GlobalRegistryModel s_model;
 
-    static GlobalRegistryCollection() {
-        s_store = WorldDatabase.Instance.Store;
-    }
+    static GlobalRegistryCollection() => s_store = WorldDatabase.Instance.Store;
 
     /// <summary>
     /// Saves a new global registry to the database.
@@ -86,4 +87,56 @@ public static class GlobalRegistryCollection {
 
         return s_model.GlobalRegistryValues[entry];
     }
+
+    /// <summary>
+    /// Checks if the global registry requirements are met.
+    /// </summary>
+    /// <param name="values"> The requirements to check. </param>
+    /// <param name="operatorType"> The operator type to use. </param>
+    /// <returns> True if all requirements are met, false otherwise. </returns>
+    public static bool CheckGlobalRegistryRequirements(List<Requirement> values, Requirement.Operator operatorType) {
+        var allMatched = true;
+
+        foreach (var requirement in values) {
+            if (requirement is ReqGlobalRegistryValue globalReq) {
+                if (!GlobalRegistryValueMet(globalReq)
+                    && operatorType == Requirement.Operator.ROP_AND) {
+                    return false;
+                }
+
+                allMatched = allMatched && !globalReq.m_applyNOT;
+            }
+            else {
+                Logger.Warning("Holy!!! We found a spawn requirement that isn't a global registry value. " +
+                            "This is a problem. Let Jooty know.");
+            }
+        }
+
+        return allMatched;
+    }
+
+    private static bool GlobalRegistryValueMet(ReqGlobalRegistryValue value) {
+        var globalValue = GetRegistryEntry(value.m_entryName);
+
+        switch (value.m_operatorType) {
+            case ReqNumeric.OPERATOR_TYPE.OPERATOR_EQUALS:
+                return value.m_numericValue == globalValue;
+            case ReqNumeric.OPERATOR_TYPE.OPERATOR_LESS_THAN:
+                return value.m_numericValue < globalValue;
+            case ReqNumeric.OPERATOR_TYPE.OPERATOR_LESS_THAN_EQ:
+                return value.m_numericValue <= globalValue;
+            case ReqNumeric.OPERATOR_TYPE.OPERATOR_GREATER_THAN:
+                return value.m_numericValue > globalValue;
+            case ReqNumeric.OPERATOR_TYPE.OPERATOR_GREATER_THAN_EQ:
+                return value.m_numericValue >= globalValue;
+            case ReqNumeric.OPERATOR_TYPE.OPERATOR_UNKNOWN:
+            default: {
+                    Logger.Error("Zone contains a spawn requirement that " +
+                                      "references a global registry value that does not exist. " +
+                                      "Entry name: {EntryName}", Logger.Args(value.m_entryName));
+                    return false;
+                }
+        }
+    }
+
 }
